+++ History.md
@@ -1,3 +1,9 @@
+### Unreleased
+
+**Minor Changes**
+
+* Add `simplify_polygon_hull` method to the CAPI factory (@oleksii-leonov) [#366](https://github.com/rgeo/rgeo/pull/366)
+
 ### 3.0.1 / 2023-11-15
 
 **Minor Changes**